requirements.txt
# ===========================================
# POLIOM HR ASSISTANT REQUIREMENTS
# ===========================================

# ===== CORE FRAMEWORKS =====
fastapi==0.104.1
uvicorn==0.24.0
aiogram==3.3.0

# ===== DATABASE =====
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
asyncpg==0.29.0
alembic==1.13.1

# ===== REDIS & CACHING =====
redis==5.0.1

# ===== CELERY (TASK QUEUE) =====
celery==5.3.4
eventlet==0.33.3
gevent==23.9.1

# ===== WEB & HTTP =====
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0
python-multipart==0.0.6
aiofiles>=23.2.1

# ===== TEMPLATES & UI =====
jinja2==3.1.2

# ===== DOCUMENT PROCESSING =====
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27

# ===== ML/AI для RAG =====
sentence-transformers==2.7.0
transformers==4.36.0
huggingface_hub==0.19.4
numpy==1.24.3
scikit-learn==1.3.2
faiss-cpu==1.7.4

# ===== TEXT PROCESSING =====
nltk==3.8.1
pymorphy2==0.9.1

# ===== AUTHENTICATION & SECURITY =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ===== CONFIGURATION =====
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ===== LOGGING =====
loguru==0.7.2

# ===== DEVELOPMENT & TESTING =====
pytest==7.4.3
pytest-asyncio==0.21.1 